agent: Move {billing,indexedstore}.go into new pkg/agent/billing (#309)
