+++ pkg/agent/billing/billing.go
@@ -1,4 +1,4 @@
-package agent
+package billing
 
 import (
 	"context"
